Refactor setOverlay into buildOverlayState module-level function
- Extracted the extremely long logic from `setOverlay` in `WebGLRenderer.tsx` into a new `buildOverlayState` helper function.
- Created `src/components/Plot/buildOverlay.ts` to host the pure sub-operation.
- Maintained exact logic and type signatures to ensure no regressions and eliminate component scope pollution.
